requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
dnspython==1.11.1
ipdb==0.8
ipython==2.1.0
itsdangerous==0.24
requests==2.3.0
wsgiref==0.1.2
